리팩토링: RAG 쿼리를 ManualChunkRetrievalRepository로 분리
service에서 JdbcTemplate 직접 사용을 제거하고 infra 레이어 @Repository로 이동. SET LOCAL 정합성을 위해 키워드 쿼리에 @Transactional(MANDATORY) 가드.
