+++ src/main/java/com/aicsassistant/analysis/dto/ScoredManualChunk.java
@@ -1,0 +1,4 @@
+package com.aicsassistant.analysis.dto;
+
+public record ScoredManualChunk(RetrievedManualChunkDto chunk, double score) {
+}
